microservices/carriers_api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
google-cloud-storage>=2.0.0
fsspec>=2025.2.0
gcsfs>=2025.2.0
google-cloud-secretmanager>=2.23.0